Add files via upload
These are updates to the training data files.  The songs Happy Birthday and Yankee Doodle were removed from the train data and will be added to the test data.  There are now more songs added to the train data for minor keys.  The SourceCode.R file contains functions that are used in the training data files.